+++ dlls/rpcrt4/rpc_defs.h
@@ -31,13 +31,13 @@ typedef struct
   unsigned char drep[4];          /* Data representation */
   unsigned short frag_len;        /* Data size in bytes including header and tail. */
   unsigned short auth_len;        /* Authentication length  */
-  unsigned long call_id;          /* Call identifier. */
+  unsigned int  call_id;          /* Call identifier. */
 } RpcPktCommonHdr;
 
 typedef struct
 {
   RpcPktCommonHdr common;
-  unsigned long alloc_hint;       /* Data size in bytes excluding header and tail. */
+  unsigned int   alloc_hint;      /* Data size in bytes excluding header and tail. */
   unsigned short context_id;      /* Presentation context identifier */
   unsigned short opnum;
 } RpcPktRequestHdr;
@@ -45,7 +45,7 @@ typedef struct
 typedef struct
 {
   RpcPktCommonHdr common;
-  unsigned long alloc_hint;       /* Data size in bytes excluding header and tail. */
+  unsigned int   alloc_hint;      /* Data size in bytes excluding header and tail. */
   unsigned short context_id;      /* Presentation context identifier */
   unsigned char cancel_count;
   unsigned char reserved;
@@ -54,12 +54,12 @@ typedef struct
 typedef struct
 {
   RpcPktCommonHdr common;
-  unsigned long alloc_hint;       /* Data size in bytes excluding header and tail. */
+  unsigned int   alloc_hint;      /* Data size in bytes excluding header and tail. */
   unsigned short context_id;      /* Presentation context identifier */
   unsigned char cancel_count;     /* Received cancel count */
   unsigned char reserved;         /* Force alignment! */
-  unsigned long status;           /* Runtime fault code (RPC_STATUS) */
-  unsigned long reserved2;
+  unsigned int  status;           /* Runtime fault code (RPC_STATUS) */
+  unsigned int  reserved2;
 } RpcPktFaultHdr;
 
 typedef struct
@@ -67,7 +67,7 @@ typedef struct
   RpcPktCommonHdr common;
   unsigned short max_tsize;       /* Maximum transmission fragment size */
   unsigned short max_rsize;       /* Maximum receive fragment size */
-  unsigned long assoc_gid;        /* Associated group id */
+  unsigned int  assoc_gid;        /* Associated group id */
   unsigned char num_elements;     /* Number of elements */
   unsigned char padding[3];       /* Force alignment! */
   unsigned short context_id;      /* Presentation context identifier */
@@ -99,7 +99,7 @@ typedef struct
   RpcPktCommonHdr common;
   unsigned short max_tsize;       /* Maximum transmission fragment size */
   unsigned short max_rsize;       /* Maximum receive fragment size */
-  unsigned long assoc_gid;        /* Associated group id */
+  unsigned int assoc_gid;         /* Associated group id */
   /* 
    * Following this header are these fields:
    *   RpcAddressString server_address;
@@ -147,7 +147,7 @@ typedef struct
   unsigned char auth_level;      /* RPC_C_AUTHN_LEVEL* */
   unsigned char auth_pad_length; /* length of padding to restore n % 4 alignment */
   unsigned char auth_reserved;   /* reserved, must be zero */
-  unsigned long auth_context_id; /* unique value for the authenticated connection */
+  unsigned int  auth_context_id; /* unique value for the authenticated connection */
 } RpcAuthVerifier;
 
 #define RPC_AUTH_VERIFIER_LEN(common_hdr) \
